+++ control-core/src/main/kotlin/at/florianschuster/control/defaultTag.kt
@@ -1,8 +1,5 @@
 package at.florianschuster.control
 
-/**
- * Creates a default tag that a [ControllerImplementation] uses for debugging.
- */
 @Suppress("NOTHING_TO_INLINE")
 internal inline fun defaultTag(): String {
     val stackTrace = Throwable().stackTrace