+++ control-core/src/test/kotlin/at/florianschuster/control/TestHelper.kt
@@ -1,0 +1,8 @@
+package at.florianschuster.control
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.CoroutineScope
+import kotlin.coroutines.ContinuationInterceptor
+
+internal val CoroutineScope.scopeDispatcher: CoroutineDispatcher
+    get() = coroutineContext[ContinuationInterceptor] as CoroutineDispatcher