+++ examples/android-counter-compose/src/androidTest/kotlin/at/florianschuster/control/androidcountercomposeexample/CounterScreenTest.kt
@@ -1,17 +1,12 @@
 package at.florianschuster.control.androidcountercomposeexample
 
-import androidx.compose.getValue
 import androidx.ui.test.assertIsDisplayed
 import androidx.ui.test.createComposeRule
 import androidx.ui.test.doClick
 import androidx.ui.test.findByTag
 import androidx.ui.test.findByText
-import at.florianschuster.control.kotlincounter.CounterAction
-import at.florianschuster.control.kotlincounter.CounterController
-import at.florianschuster.control.kotlincounter.CounterState
-import at.florianschuster.control.kotlincounter.createCounterController
+import at.florianschuster.control.ControllerStub
 import at.florianschuster.control.stub
-import kotlinx.coroutines.test.TestCoroutineScope
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -25,15 +20,13 @@ internal class CounterScreenTest {
     @get:Rule
     val composeTestRule = createComposeRule()
 
-    private lateinit var controller: CounterController
+    private lateinit var stub: ControllerStub<CounterAction, CounterState>
 
     @Before
     fun setup() {
-        controller = TestCoroutineScope().createCounterController().apply { stub() }
-        composeTestRule.setContent {
-            val state by controller.collectState()
-            CounterScreen(counterState = state, action = controller::dispatch)
-        }
+        val controller = CounterController()
+        stub = controller.stub()
+        composeTestRule.setContent { CounterScreen(controller) }
     }
 
     @Test
@@ -45,7 +38,7 @@ internal class CounterScreenTest {
         }
 
         // then
-        assertEquals(CounterAction.Increment, controller.stub().dispatchedActions.last())
+        assertEquals(CounterAction.Increment, stub.dispatchedActions.last())
     }
 
     @Test
@@ -57,7 +50,7 @@ internal class CounterScreenTest {
         }
 
         // then
-        assertEquals(CounterAction.Decrement, controller.stub().dispatchedActions.last())
+        assertEquals(CounterAction.Decrement, stub.dispatchedActions.last())
     }
 
     @Test
@@ -66,7 +59,7 @@ internal class CounterScreenTest {
         val testValue = 42
 
         // when
-        controller.stub().emitState(CounterState(value = testValue))
+        stub.emitState(CounterState(value = testValue))
 
         // then
         findByText("$testValue").assertIsDisplayed()
@@ -75,7 +68,7 @@ internal class CounterScreenTest {
     @Test
     fun whenStateOffersNotLoadingProgressBarDoesNotExist() {
         // when
-        controller.stub().emitState(CounterState(loading = false))
+        stub.emitState(CounterState(loading = false))
 
         // then
         findByTag("progressIndicator").assertDoesNotExist()