+++ examples/android-counter-compose/src/main/kotlin/at/florianschuster/control/androidcountercomposeexample/CounterScreen.kt
@@ -1,6 +1,8 @@
 package at.florianschuster.control.androidcountercomposeexample
 
 import androidx.compose.Composable
+import androidx.compose.getValue
+import androidx.compose.remember
 import androidx.ui.core.Alignment
 import androidx.ui.core.Modifier
 import androidx.ui.core.tag
@@ -16,20 +18,27 @@ import androidx.ui.material.MaterialTheme
 import androidx.ui.material.TextButton
 import androidx.ui.tooling.preview.Preview
 import androidx.ui.unit.dp
-import at.florianschuster.control.kotlincounter.CounterAction
-import at.florianschuster.control.kotlincounter.CounterState
 
 @Composable
 internal fun CounterScreen(
+    injectedController: CounterController = CounterController()
+) {
+    val controller = remember { injectedController }
+    val counterState by controller.collectState()
+    CounterComponent(counterState, controller::dispatch)
+}
+
+@Composable
+private fun CounterComponent(
     counterState: CounterState,
-    action: (CounterAction) -> Unit = {}
+    dispatch: (CounterAction) -> Unit = {}
 ) {
     Stack(modifier = Modifier.fillMaxSize()) {
         Row(
             modifier = Modifier.fillMaxWidth().gravity(Alignment.Center),
             horizontalArrangement = Arrangement.SpaceEvenly
         ) {
-            TextButton(onClick = { action(CounterAction.Decrement) }) {
+            TextButton(onClick = { dispatch(CounterAction.Decrement) }) {
                 Text(text = "-", style = MaterialTheme.typography.h4)
             }
             Text(
@@ -38,7 +47,7 @@ internal fun CounterScreen(
                 style = MaterialTheme.typography.h3,
                 modifier = Modifier.tag("valueText")
             )
-            TextButton(onClick = { action(CounterAction.Increment) }) {
+            TextButton(onClick = { dispatch(CounterAction.Increment) }) {
                 Text(text = "+", style = MaterialTheme.typography.h4)
             }
         }
@@ -53,18 +62,18 @@ internal fun CounterScreen(
     }
 }
 
-@Preview(name = "Loading")
+@Preview("not loading")
 @Composable
-private fun CounterScreenPreviewLoading() {
+private fun CounterComponentPreview() {
     MaterialTheme(colors = AppColors.currentColorPalette) {
-        CounterScreen(counterState = CounterState(value = 21, loading = true))
+        CounterComponent(CounterState(value = -1, loading = false))
     }
 }
 
-@Preview(name = "Not Loading")
+@Preview("loading")
 @Composable
-private fun CounterScreenPreviewNotLoading() {
+private fun CounterComponentPreviewLoading() {
     MaterialTheme(colors = AppColors.currentColorPalette) {
-        CounterScreen(counterState = CounterState(value = 21, loading = false))
+        CounterComponent(CounterState(value = 1, loading = true))
     }
 }