+++ examples/android-counter/src/androidTest/kotlin/at/florianschuster/control/androidcounter/CounterViewTest.kt
@@ -8,8 +8,8 @@ import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import at.florianschuster.control.ControllerStub
 import at.florianschuster.control.kotlincounter.CounterAction
-import at.florianschuster.control.kotlincounter.CounterController
 import at.florianschuster.control.kotlincounter.CounterState
 import at.florianschuster.control.kotlincounter.createCounterController
 import at.florianschuster.control.stub
@@ -21,12 +21,13 @@ import kotlin.test.assertEquals
 @RunWith(AndroidJUnit4::class)
 internal class CounterViewTest {
 
-    private lateinit var controller: CounterController
+    private lateinit var stub: ControllerStub<CounterAction, CounterState>
 
     @Before
     fun setup() {
         CounterView.CounterControllerProvider = { scope ->
-            controller = scope.createCounterController().apply { stub() }
+            val controller = scope.createCounterController()
+            stub = controller.stub()
             controller
         }
         launchFragmentInContainer<CounterView>()
@@ -38,7 +39,7 @@ internal class CounterViewTest {
         onView(withId(R.id.increaseButton)).perform(click())
 
         // then
-        assertEquals(CounterAction.Increment, controller.stub().dispatchedActions.last())
+        assertEquals(CounterAction.Increment, stub.dispatchedActions.last())
     }
 
     @Test
@@ -47,7 +48,7 @@ internal class CounterViewTest {
         onView(withId(R.id.decreaseButton)).perform(click())
 
         // then
-        assertEquals(CounterAction.Decrement, controller.stub().dispatchedActions.last())
+        assertEquals(CounterAction.Decrement, stub.dispatchedActions.last())
     }
 
     @Test
@@ -56,7 +57,7 @@ internal class CounterViewTest {
         val testValue = 1
 
         // when
-        controller.stub().emitState(CounterState(value = testValue))
+        stub.emitState(CounterState(value = testValue))
 
         // then
         onView(withId(R.id.valueTextView)).check(matches(withText("$testValue")))
@@ -65,7 +66,7 @@ internal class CounterViewTest {
     @Test
     fun whenStateOffersLoadingProgressBarIsVisible() {
         // when
-        controller.stub().emitState(CounterState(loading = true))
+        stub.emitState(CounterState(loading = true))
 
         // then
         onView(withId(R.id.loadingProgressBar)).check(matches(isDisplayed()))