+++ examples/android-github/src/main/kotlin/at/florianschuster/control/androidgithub/MainActivity.kt
@@ -2,16 +2,17 @@ package at.florianschuster.control.androidgithub
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.commit
 import at.florianschuster.control.androidgithub.search.GithubView
 
 internal class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        if (savedInstanceState != null) return
-        supportFragmentManager.beginTransaction()
-            .replace(android.R.id.content, GithubView())
-            .commit()
+        if (savedInstanceState == null) {
+            supportFragmentManager.commit {
+                replace(android.R.id.content, GithubView())
+            }
+        }
     }
 }