+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # changelog
 
-## `[X.X.X]` - Unreleased
+## `[1.0.0]` - Unreleased
 
+- binary compatibility will now be verified on every release.
+
+## `[0.11.0]` - 2020-05-30
+
+- `CoroutineScope.createController` and `CoroutineScope.createSynchronousController` now accept a custom `ControllerStart` parameter instead of `CoroutineStart`.
+- Add `ManagedController`.
+- `Controller.stub` is now marked as `@TestOnly`.
 - binary compatibility is now verified on each `[build]` & `[publish]`.
-- `Controller.stub` is now marked as `@TestOnly`
 
 ## `[0.10.0]` - 2020-05-11
 
@@ -12,5 +18,5 @@
 
 ## `[0.9.0]` - 2020-05-10
 
-- `ControllerImplemenation` now uses `MutableStateFlow` instead of `ConflatedBroadCastChannel` internally.
+- `ControllerImplementation` now uses `MutableStateFlow` instead of `ConflatedBroadCastChannel` internally.
 - `Controller.state` emissions are now distinct by default (via `StateFlow`).